tried adding community finder
i tried implementing the community finder function for 10 communities & looking at states of nodes that are in the largest of the 10 communities at some time t = 14. not sure how to go further (i.e. removing edges that connected a community to the rest of the graph, and doing that at a specific point in time?).